Enum functions -> Static to Extern
Since the functions are not part of a class there is no reason for them to be static. They just need to be available in the paces where the header Enum will be used.
The definition is not included in the .h file and they are not part of a class to make an instance of so extern is the most effective option.

Furthermore, there are multiple methods for conversion including maps and operator overload. Class with std::map parent polymorphism. Alternative is template classes or functions. The number of conversions is small so  simple function overloading is enough to cover all cases without additional complexity.